Add unsubscribe list
To setup unsubscribes, configure the database url, then run
$ createdb cryptobot
$ ./unsubscribe --create

To unsubscribe an address
$ ./unsubscribe --add foo@example.com

To un-unsubscribe an address...(TODO)

Unsubscribed emails are salted and hashed before storage. The salt is a
random string generated during database creation, then stored in the
database itself. This makes it harder for the operator to shoot
themselves in the foot, by say, modifying a param in a config file.